Move get_sensors and load_current_sensor_state from thermostat.py to zone.py
