+++ nexia/zone.py
@@ -1,6 +1,8 @@
-"""Nexia Themostat Zone."""
+"""Nexia Thermostat Zone."""
 from __future__ import annotations
 
+import asyncio
+import json
 import logging
 import math
 from typing import TYPE_CHECKING, Any
@@ -17,6 +19,7 @@ from .const import (
     UNIT_CELSIUS,
     ZONE_IDLE,
 )
+from .sensor import NexiaSensor
 from .util import find_dict_with_keyvalue_in_json
 
 _LOGGER = logging.getLogger(__name__)
@@ -251,6 +254,23 @@ class NexiaThermostatZone:
             return True
         return run_mode["current_value"] == HOLD_PERMANENT
 
+    def get_sensors(self) -> list[NexiaSensor]:
+        """Get the sensor detail data objects from this instance
+        :return: list of sensor detail data objects
+        """
+        sensors: list[NexiaSensor] = []
+        try:
+            sensors_json = self._get_zone_features("room_iq_sensors")["sensors"]
+
+            for sensor_json in sensors_json:
+                sensors.append(NexiaSensor(*[sensor_json[fld] for fld in NexiaSensor._fields]))
+
+        except KeyError:
+            # our json has no sensors
+            pass
+
+        return sensors
+
     ########################################################################
     # Zone Set Methods
 
@@ -428,9 +448,43 @@ class NexiaThermostatZone:
                 f"{OPERATION_MODES}"
             )
 
+    async def load_current_sensor_state(self, polling_delay = 0.7, max_polls = 50) -> bool:
+        """Load the current state of its sensors into the physical thermostat.
+        :param polling_delay: seconds to wait before each polling attempt
+        :param max_polls: maximum number of times to poll for completion
+        :return: bool indicating completed
+        """
+        req_cur_state = self.API_MOBILE_ZONE_URL.format(
+            end_point="request_current_sensor_state", zone_id=self.zone_id)
+
+        async with await self._nexia_home.post_url(req_cur_state, {}) as response:
+            # The polling path in the response has the form:
+            #   https://www.mynexia.com/backstage/announcements/<48-hex-digits>
+            polling_url = self._nexia_home.resolve_url(
+                (await response.json())["result"]["polling_path"])
+        retries = max_polls
+
+        while retries:
+            await asyncio.sleep(polling_delay)
+            async with await self._nexia_home.get_url(polling_url) as response:
+                payload = (await response.read()).strip()
+
+            if payload != b"null":
+                status = json.loads(payload)["status"]
+
+                if status != "success":
+                    _LOGGER.error("Unexpected status [%s] loading current sensor state",
+                                  status)
+                return True
+            retries -= 1
+        # end while waiting for status
+
+        _LOGGER.error("Gave up waiting for current sensor state")
+        return False
+
     def round_temp(self, temperature: float) -> float:
         """
-        Rounds the temperature to the nearest 1/2 degree for C and neareast 1
+        Rounds the temperature to the nearest 1/2 degree for C and nearest 1
         degree for F
         :param temperature: temperature to round
         :return: float rounded temperature
@@ -471,7 +525,7 @@ class NexiaThermostatZone:
 
     def _get_zone_setting_or_none(self, key: str) -> Any:
         """
-        Returns the zone value from the provided key in the zones's
+        Returns the zone value from the provided key in the zone's
         JSON.
         :param key: str
         :return: value
